+++ .gitignore
@@ -5,3 +5,4 @@ yarn-error.log
 .DS_Store
 
 .now
+serviceAccountKey.json
