+++ package.json
@@ -18,7 +18,7 @@
     "add": "^2.0.6",
     "classnames": "^2.2.6",
     "dotenv-webpack": "^1.7.0",
-    "firebase": "^6.3.4",
+    "firebase": "^7.6.1",
     "fireworks-canvas": "^2.4.0",
     "husky": "^4.2.3",
     "i18next": "^19.4.4",
