+++ src/lib/api.ts
@@ -1,10 +1,7 @@
-import { play } from "~/lib/ai";
-import { updateGame } from "~/lib/firebase";
 import { ID, uniqueId } from "~/lib/id";
-import { sleep } from "~/lib/promise";
-import IGameState, { IGameStatus } from "~/lib/state";
+import IGameState from "~/lib/state";
 
-export async function api(path: string, payload: any) {
+export async function sendRequest(path: string, payload: any) {
   return fetch(path, {
     method: "POST",
     body: JSON.stringify(payload),
@@ -25,28 +22,4 @@ export async function getPlayerIdFromSession(req): Promise<ID> {
   }
 
   return playerId;
-}
-
-export async function playForBots(game: IGameState) {
-  while (true) {
-    const nextPlayer = game.players[game.currentPlayer];
-    if (!nextPlayer.bot) {
-      break;
-    }
-    if (game.status === IGameStatus.OVER) {
-      break;
-    }
-
-    if (game.options.botsWait) {
-      nextPlayer.reaction = "🧠";
-      await updateGame(game);
-    }
-
-    await sleep(game.options.botsWait);
-
-    game = play(game);
-    game.players.find(player => player.id === nextPlayer.id).reaction = null;
-
-    await updateGame(game);
-  }
 }
